Removing server files used for testing.
